cloud_requirements.txt
tensorflow==2.15.0
uvicorn==0.25.0
fastapi==0.108.0
python-multipart==0.0.6
python-dotenv==1.0.0
